fix inability to bind Control or Command
the hook failed to take into account that if `e.key`
is 'Alt', 'Control', 'Meta', or 'Shift',
the respective modifier property on
the Keyboard event will also be true.
this changes `parseKeyString` to set
the modifier flags correctly if the key is a modifier.
also adds a note to the README to use
'Control' or 'Meta' to add binds to those keys